feat: separates LeftPanelInfo code into component
